Update file structure in concordance with main
