--- dna_diffusion/metrics/sequence_quality_assessment/deopen/requirements.txt
@@ -1,4 +1,0 @@
-h5py
-hickle
-Scikit-learn==0.18.2
-nolearn==0.6.0
