--- dna_diffusion/metrics/sequence_quality_assessment/enformer/requirements.txt
@@ -1,6 +1,0 @@
-dm-sonnet==2.0.0
-kipoiseq==0.5.2
-numpy==1.19.5
-pandas==1.2.3
-tensorflow==2.4.1
-tensorflow-hub==0.11.0
