--- models/vanilla_diffusion/README.MD
@@ -1,5 +1,0 @@
-
-Current version
----
-V2 
-  - Code_to_refactor_UNET_ANNOTATED v2.ipynb
