--- models/vanilla_diffusion/archived_versions/Readme.md
@@ -1,1 +1,0 @@
-## Archived versions
